refactor: use `String` for return when possible
